+++ CHANGELOG.md
@@ -230,6 +230,10 @@ The present file will list all changes made to the project; according to the
 - The `$target` parameter has been removed from the `Rule::showRulePreviewCriteriasForm()`, `Rule::showRulePreviewResultsForm()`, `RuleCollection::showRulesEnginePreviewCriteriasForm()`, and `RuleCollection::showRulesEnginePreviewResultsForm()` methods signature.
 - `Hooks::SHOW_IN_TIMELINE`/`show_in_timeline` plugin hook has been renamed to `Hooks::TIMELINE_ITEMS`/`timeline_items`.
 - `Auth::getMethodName()` now only returns the name without a link. Use `Auth::getMethodLink()` to get a HTML-safe link.
+- Loading of TinyMCE via `Html::requireJs()` should only be used if there is unique requirements to initialize an editor.
+  `Html::requireJs('tinymce')` now only loads the main `tinymce.js` library script and should be avoided when possible.
+  Any usage of rich text editors through `Html::textarea()` or the `textareaField` macro in `templates/components/form/fields_macros.html.twig` will automatically load the TinyMCE library as needed.
+  Otherwise, you can import the `js/modules/Form/TinyMCEEditor.js` script directly.
 
 #### Deprecated
 - Usage of the `/marketplace` path for plugins URLs. All plugins URLs should now start with `/plugins`.